feat: Make _ToggleButton public
