+++ lib/components/toggle_button.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class ToggleButton extends StatelessWidget {
+  const ToggleButton({
+    super.key,
+    required this.text,
+    required this.value,
+    required this.onToggled,
+  });
+
+  final String text;
+  final bool value;
+  final ValueChanged<bool> onToggled;
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlinedButton(
+      style: OutlinedButton.styleFrom(
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+      ),
+      child: Text(
+        '$text: $value',
+        style: const TextStyle(
+          fontSize: 16,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+      onPressed: () => onToggled(!value),
+    );
+  }
+}
